ir_wall was updated in order to calculate properly the values of the angle
there is now transtion between forward to following walls. This is done because before when we were in following the wall we stopped following the walls after !walls detected. The problem was that we were not able to come back. All these problems will change after the brain is implemented.